⚡️(templates) remove transparency when not necessary to convert to jpg
Easy thumbnails converts images to jpg when there is no transparency,
but keeps png otherwise. We can't prevent users from uploading images
with transparency in places where it is not desired but we can remove
it when generating thumbnails so that the image can be converted to
jpg and thus be much smaller.