🚸(courses) include new pages in navigation by default for breadcrumbs
When creating a new page from the DjangoCMS page wizard, we did not set
the "in_navigation" flag to "True" by default. This resulted in the new
page not being present in the breadcrumb. Users had to systematically
tick the flag to fix that. We should set this flag to "True" by default.