💄(course-glimpse) fix layout when organization & course titles are long
On course glimpse, course and organization titles was not truncated when they
were too long. So when these two contents were too long, they overlap. So we
decided to truncate course title if it exceeds 3 lines and organization title if
it exceeds 2 lines. Furthermore, we decrease font-size of organization title to
0.7rem to homogenize font size within course glimpse.

Fix #1560