📌(frontend) pin @types/faker to version 5.5.9
Since faker has been take over from the community, types are integrating
directly into the new project. As the version integrating types are
still an alpha we do not want yet switch to this project. In the
meantime, a new version of @types/faker has been released which removed
all the types. As we are still using the "marak" faker project, we need
to pin @types/faker.