♻️(pagination) simplify django/react components a11y labels handling
there is no need to have an ID and an `aria-labelledby` attribute. We
can have the exact same behavior by directly passing a label via the
`aria-label` attribute.

This simplifies the code a tiny code, removing the need to handle
different ids when using the pagination components several times on one
page.