🚸(course) add autocomplete feature on Page select in Person plugin
Some sites can have a huge list of objects where it is hard to search for a
specific object. The page selector from Person plugin is one of this case.

We implement usage of an autocomplete feature with Select2.js by the way of
new dependancy 'django-autocomplete-light' enabled and configured on the
'page' field from Person plugin form.