♿️(pagination) improve a11y labels in django templates
When using multiple paginations on the same page, we need to have
different labels on each pagination.

This is mostly because a screen-reader user can, via keyboard shortcut,
jump directly to each <nav> element. So if each pagination <nav> is
labelled by "Pagination", it isn't helpful.

The dj-pagination {% paginate %} template tag doesn't allow us to pass
custom variables, so we wrap it to pass it via {% include %} tags.