💄(layout) fix usage of deprecated Sass division operator
Since the project use Dart Sass to build CSS from Sass, there was some
deprecation warning messages about the division operator character "/" which
is deprecated in favor of 'math.div'.

This is solid futur proof Sass reference so there is no need to wait anymore
since there was not many usage of it in our Sass sources.

https://sass-lang.com/documentation/breaking-changes/slash-div