🐛(course) fix cms toolbar bug when editing a plugin from a detail page
It has been reported than when editing a plugin a bug often appear which
turn the CMS toolbar in mobile mode.

It was due to usage of 'get_placeholder_plugins' and 'blockplugin' tag in page
metas. This is because in edit mode these tags append some HTML required for
the CMS interface to edit in place a plugin and so it append some HTML
elements like 'div' and 'template' which are totally invalid in '<head>'
elements and also some Javascript code which invoke some object/function that
are not yet loaded.

This commit adds a new keyword argument 'silent' to previously cited template
tags. This argument when set to 'True' (string) disable default behavior of
edition mode and so does not inject any markup related to placeholder and
plugin edition. And finally, all detail template now use this argument on tags
inside '<head>'.