✨(courses) add 'excerpt' placeholder to Organization detail template
Organization plugin 'row' variant has been reported to have broken layout
because of wrong CSS and the description placeholder which can include links.

It leaded to decision to add a new placeholder 'excerpt' alike with Blogpost,
so Organization 'row' variant can still be in a link container since its text
content is always plain text. The excerpt is used if not empty, else it will
fallback to use the description but with every markup stripped out.