🐛(course) fix course run deletion when translation title is empty
It has been reported than trying to delete a course run fails to an error
when a course run translation have an empty title.

This changes the '__str__' from CourseRunTranslation model to a fallback title
"Empty title" when its title is empty. CourseRun deletion won't fail again
because of empty translation title.