✨(core) add site name in <title> everywhere
By default, the <title> only tells us about the current page; it does
not tell us about what website we are on.

Adding the site name after the page name mostly improves two things:

- accessibility. Knowing what is the current site is a great
improvement. Without it, if I have multiple tabs opened, all related to
learning stuff, and a few tabs don't have their website name in the
title, it can be cumbersome to navigate with a screen-reader and
remember what tab is what website.
- SEO