♿️(course) make the course heading the first thing screen readers announce
Since the badges are relative to the course, we want them to be read
*after* the course heading.

Headings are a way to navigate inside the page when using a screen
reader. I can jump from heading to heading to parse content.
So if I have some content related to the section described by this
heading, that is written *before* the heading, I can easily miss it.

Now I can't!