♿️(course glimpse) put the course icon after the heading
putting info related to the course glimpse before its heading is not
great for two reasons:

- with a screen reader, when navigate by headings on the page to jump
between each course glimpse, it makes users miss the info
- with a screen reader, when navigating in the content "normally" (with
down arrow), the first thing announced for a glimpse course is the icon
title, which is pretty misleading. We should hear the course name
instead.