💄(card) add a "stretched link" feature on cards
Idea and implementation taken from bootstrap (thanks @sveetch).

The idea is to add the possibility to have a __stretched-link <a> inside
a card, that automatically makes the link take its whole container as a
clickable zone.

Note that since this works in relation to the nearest position: relative
parent, in our case we will usually need to have multiple links (one for
media and one for content).

This is useful because we want to avoid having cards that are big <a>
containing everything, as doing this makes it quickly annoying, or even
sometimes pretty hard to understand with a screen reader.