♿️(program) restructure a program glimpse link for better screen reader UX
- replace the <a> wrapping everything by a div
- add a <a> around the media that is only for mouse users
- add a stretched-link <a> in the glimpse title. This one is accessible
to keyboard/screen reader users, and is clickable in the whole content
zone for mouse users

The behavior is the same as before for mouse users.

This makes the glimpse clearer when using a screen reader, because now
only the actual link is announced as such. Before, depending on screen
reader software/config, it could announce the whole title+content as the
link in one block, can be quite difficult to understand.