♿️(styles) improve the offscreen class
Change the implementation of the offscreen class to match the one from
bootstrap that is battle-tested.

The main idea is to stop using "left: -100000px" that can cause
scrolling "flicker" or weird focus styles when using a screen reader.
This can be annoying for sighted people that use a screen reader.