♿️(course glimpse) put the <a> only around media and title
this makes for a better UX for screen reader/keyboard users.

The idea is to have a big clickable zone for mouse users with the link
on the media cover + on the title, with the link on the title having a
bigger clickable space than shown.

We manually add/remove the card shadow when the link is hovered/focused

src/richie/apps/courses/templates/courses/cms/fragment_course_glimpse.html