♿️(blogpost glimpse) improve screen reader UX with the date
before, the date was announced by screen reader before the title.

That's problematic for 2 reasons:

- when navigating in the page from top to bottom, we first hear the
date, then the title. We should hear the title first, to understand the
what the date is about.
- when navigating in the page from title to title, we can miss the date,
or believe the date concerning the next blogpost is about the current
one in our navigation

Duplicating the date in the correct DOM location only for screen
readers. Because it was too complicated to show the date above the cover
*and* announce it at the right time for screen readers.