♿️(templates) move a few DOM elements for better screen reader UX
In the blogpost page and blogpost fragments, the categories/date
/author were announced by a screen reader before the title.

Make it so that they are announced after the title, to understand
the context better, without moving them visually.

We do this because headings are a way to navigate inside the
page when using a screen reader. I can jump from heading to heading
to parse content. So if I have some content related to the section
described by this heading, that is written *before* the heading,
I can easily miss it. Now I can't miss it.