♿️(templates) label categories for better screen reader UX
In person glimpses and blogpost glimpses, blogpost list, etc. it's
not that obvious that the category tags are categories.

So, at least for screen readers, add a label.

I guess the better thing to do would be to show this label in the UI
too, but it's open to debate. We can argue the "tag" visuals help us
understand enough.