♿️(course) improve a11y of the organization image link
On a course page, the organization image link had the organization name
as alternative text. That's great but we can do better. By reading only
its name, we don't really understand that it is the _organization_ name.
Be more explicit by adding a label before the name.

We also make sure the alt text is announced only once by screen reader
by forcing the aria-label and removing the image alt, that becomes
unnecessary