📝(website) rework quickstart and installation documentation
We should better differentiate what is discovering Richie (visiting the
demo site), starting your own site with Richie (spawning a site factory
with Cookiecutter) and contributing to Richie (bootstrapping the docker
stack or making a native installation). Each objective is a totally
different approach. We should present them one by one starting from the
easiest and more common to the hardest and rare implication of developpers.