📝(website) fix and improve website documentation
The website was not building anymore following our refactoring of
the documentation to include cookiecutter instructions.

There seem to be a bug on docusaurus versioning that prevents us from
renaming an existing page... so I fixed it by renaming it also in the
history. This does not impact what users see and the old versions of
the website are still available.

I also improved the documentation to make it clearer. In particular
I repeated part of the README in the first page of the documentation
to introduce the project a minimum.