♿️(templates) improve screen reader UX with changed header/nav markup
there were a few issues before:

- the <nav> tag was inside the <header> tag. That prevented NVDA from
recognizing the <nav> tag in its UI that lists all regions of the page.
Now NVDA lists both the <header> and the <nav>. The problem wasn't that
huge, and users are kinda used of this nuisance and usually end up
navigation in the <header> by hand to find the <nav>, but if we can
prevent it, it's better.
- the hamburger button was placed before the home link in the DOM. It's
now placed right before the area it toggles, so that keyboard navigation
makes sense to the user, at it follows the DOM.
- the "aside" menu was placed after the main menu in the DOM, but it
appeared before in the UI. That made keyboard navigation a bit weird
since it didn't follow the UI. It's now better placed.