♿️(search) improve a11y on filters pane with a small screen
- put the existing filters toggle button above the content it toggles in
the DOM, so that tab order makes sense. Before that, when navigating via
keyboard, we would end up in the search results right after opening the
filters pane, that made little sense.

- add a 2nd filters toggle button *after* the content it toggles in the
DOM, shown only when the pane is opened. So that when we are at the
bottom of the pane, screen reader users can easily close it, and sighted
users don't have to scroll to the top anymore.

- make sure that when the filter pane is visually hidden, it is actually
hidden from screen readers. An "aria-hidden" attribute was set, but it
wasn't enough, we could still use tab to go through form elements inside
it. Now we use "display: none" to hide it for everyone (being careful
about visual transitions), making the aria-hidden attribute unnecessary.

- for screen reader users using small screens, add region roles to the
filters and results div. In desktop mode, it's not really necessary
since the <h2> are there and we can easily navigate by heading. But in
mobile view, the filters pane <h2> is hidden. So in that case, the fact
that we can filter courses is kinda hidden. Only the search results
heading is viewed by screen readers. The button is there sure, but it's
not that "discoverable". Adding region roles (labelled by the
already-there <h2>s) now makes the filters directly noticeable even when
the filters pane is closed.