♿️(frontend) rework headings hierarchy of course product widget
- There were a few headings in the product widget that we may consider a
bit overkill, since the product widget is filled with list items, it's
not really necessary to add headings to them.

- Since we can potentially have multiple product widgets, add a hidden
heading that acts as a container of this section.

- Headings in the sale tunnel were set really low but this is not necessary.
On paper we could even use <h1>s because its in a modal. But I guess I'm used
to keeping only one <h1> by page. So we start them at <h2>, and make sure
the hierarchy has no "gap". That allows more heading levels than starting at
<h5> if necessary.