♿️(frontend) make sure SR users understand dates of product widget
for screen reader users, the problem was that things werent
announced in the correct order before and it could quickly
get difficult to understand.

For example, it said:

Course 0, Start, End, May 09 2022, Jun 25 2022

Now it says:

Course 0, Start, May 09 2022, End, Jun 25 2022

Since the UI is a bit tricky, the idea is to duplicate in offscreen
classes the start/end label, and hide the visible ones from screen
readers. Simple fix!