♿️(templates) be explicit about the type of content in the page title
It's not really explicit when we read the title that we are on a
category page or a program page.

Especially with a screen reader, the <title> is the main thing that
gives context as to what is the page. So it's pretty important to be
explicit in this case.

Maybe this is overkill but I'd rather be a bit too explicit here
than not enough.