♿️(frontend) sale tunnel address creation: make it work with keyboard
In the sale tunnel payment step:

- the address creation button was hidden from screen reader users,
changing that.

- activating the "create an address" changes the content of the modal.
Since the focused element is removed from the DOM when this happens, by
default the focused element becomes <body>. This is bad for users
navigating with keyboard/screen readers. Now we manually update the
focused element when switching from address creation form/payment info.