♿️(frontend) sale tunnel address: make <hx> consistent w/ other steps
starting the headings at h5 is not really necessary, and now other steps
begin at h2, so follow this. We could argue we would want to begin at h4
if the personal info/billing address text was still shown to the user
when the form is shown, but this is not the case.

There is no real "right" answer I guess.