♻️(settings) make filter machine name implications clearer
As we noted in issue #1350, some filter machine names are free
form (those that connect to categories) while others are constrained
(those that connect with organizations or persons).

This can create unexpected bugs if consumers of Richie operate from
the starting point of categories: that the filter machine names are
customizable at-will.

We can add a small comment in-place to warn users re-using those
settings that "organizations" and "persons" are special names that
logically connect filters with their corresponding backend objects.