💄(blogpost_detail) move categories list on the right side
Currently, categories list are above the blogpost metadata list. That is
confusing as metadata also include categories list. So we decide to move
categories list on the right side below the related posts list.

Furthermore, we add a block "content_header" to continue to be able to put
content above blogpost content.

Close #1707