♿️(frontend) better a11y on product widget enroll{able,ed} courses
in the product widget, when you can enroll in a course, there were a few
things that needed improvement for keyboard and screen reader users.

Now:

- the DOM structure is improved a bit, we removed the <li> tags that
were wrapping the action buttons, because they are not really part of
the course *list*. It's now clearer for screen reader users.

- the feedback messages when you can't enroll in a course were all in
the disabled buttons. This made them totally invisible for screen reader
users, because a disabled button is removed from the SR's list of things
that are focusable. Now, the action buttons are always clickable, and a
feedback message is shown next to it when you can't enroll. For SR
users, the feedback is focused when it appears so that it is announced
directly.

- when you enroll/unroll, keyboard navigation was disrupted because the
buttons that are just clicked are removed from the DOM to be replaced by
another button. We handle this specific case with a mutation observer so
that keyboard users/SR users don't get lost.