♿️(frontend) improve error handling on inputs
if a field has a message (error or info), this message should be
programmatically tied to its input. This is done so that screen reader
users detect in a easier way when a field has an error.

We do this by adding an aria-describedby attribute on a input when we
detect a message.