♿️(frontend) sale tunnel address: handle all keyboard navigations
- there were multiple ways to go to the address form and back to the
payment step in the sale tunnel. So instead of relying on one toggle
button that exists in only one case, now, when we go back to the payment
step from the adress form, we always focus back the adress block title.

- when we had an existing address that we wanted to modify, clicking
modify was a bit confusing for a SR user because he wasn't notified of
the form below being filled automatically. Now, after clicking the "edit
address" button, we focus the first form input