🔨(cookiecutter) enable individual configuration files out of tree files
In the `env.d` folder, we create a `development.dist` file used
as a template. The one used in the project is now generated.
Each developer will now be able to modify any key without risking
committing it to the tree files. This implied deleting current
configuration files present in the distribution