➕(easy_thumbnails) add easy_thumbnails[svg]
Since version 2.8.2, easy_thumbnails make svg support optional. If we want
support this image format we have to install easy_thumbnails with extra
dependencies. This dependency is used by django-filer but this package does not
currently install extra svg dependencies. So as a quick workaround, we add
easy_thumbnails[svg] to our requirements.