♿️(frontend) sale tunnel address: show form error messages
before this commit, the user didn't have any feedback on a form error
because he couldn't submit the form before it being valid. But how can
he know what is a "valid" form?

The better way is usually to let the user submit an incorrect form and
get the appropriate feedback instead of having to guess.

So now, the user can submit the form whenever, and error messages are
shown accordingly. react-hook-form handles focusing the first input that
is invalid after the form submission