♻️(search) refactor the way filters are being configured
It was not easy to define the list of filters to display and their order.
With this change, it can be done with a separate setting which is a simple
list. Filter definitions become a dictionary which also makes it easier as
compared with the previous list of tuples.

This requires upgrade attention only for sites that have customized their
filters.