🔨(bin) stop using env variable to activate sites
The DX was not good as we had to run bin/activate and then export
a variable. Many users forgot to export the variable expecting the
activate script to do what it claims to do...

We now use a `.site` file to record the active site as it is done
in terraform. This allows a much better DX as running `bin/activate`
now effectively activates the site.