⬆️(frontend) reset yarn.lock to sync @types/react versions
There was a type error with @types/react package as react-intl and our package
are both using this package. Since the upgrade two different versions of this
package was used and cause a duplicated declaration error. So reset yarn.lock
force react-intl to use our @types/react version as it satisfies its
requirements.