♻️(frontend) move useOrders to useResources hook
Migrate the code to take advantage of useResources hooks.
It implied a modification into PaymentButton because useResources
uses `mutate` instead of `mutateAsync` ( that useOrders was using ),
then we had to make use of the onSuccess callback, which is the only
way to get the response of the mutation when using `mutate`, just
awaiting its value always gives undefined.