✨(frontend) add DashboardItemOrder that can choose course runs
This is a big part that implied lots of updates in the codebase.
The course runs list is made encapsulated in its own logic in
DashboardItemCourseEnrolling in order to be also used by
DashboardItemEnrollment. This component is able to be readonly
or writable, meaning that it can only display the active
course run or the list of available course runs with buttons
allowing the user to switch between them.