🐛(CourseProductItem) prevent unmount SaleTunnel on order creation
OrderQuery invalidates itself on mutation success. By default all hooks
linked to this query triggered a rerender on this invalidation. About
the CourseProductItem that's problematic as this mutation is called
within the SaleTunnel, so the orderQuery fetching state switch to true then
the Spinner is show and SaleTunnel is unmounted. In order to prevent this
weird behavior, we rework the logic to display product information and
spinner.