🐛(frontend) fix CourseProductItem issue on session state update
About CourseProductItem, previously, we guessed if current user has purchased
the related product from the product response itself. Indeed, if the user has
purchased the product, the product's property `orders` contains all pending and
validated orders related to the product and the current user. But there was a
bug as useProduct hook is not aware about session state. So in the case where
the page was displayed without session state initialized, the product query was
executed immediately then when the session state was updated, as the product
query was not stale, it was not executed again.

In order to fix that, to guess if the authenticated user has purchased the
related product, we use the `useOrders` hook which is aware of session state
update.