🚸(frontend) manage pending state on CourseProductItem
In some case, user can have an order in a pending state for a product. In order
to prevent user to open sale tunnel in this case we do not show the purchase
button. But the label currently display show "Enrolled" that is unclear. In
order to improve user understanding, we decide to display "Pending" label in
place of the current "Enrolled" label in this case.