✅(frontend) remove some warnings about DashboardCourses test suite
DashboardCourses test suite triggers to warning on run. First, the simpler, we
mismocked a request, so fetchMock warns us about that. Second, it seems during
we are waiting that all assertion are resolved, a rerender is trigger and a
warning is emitted that this code block should be wrapped into an act. Wrapping
this assertion into waitFor seems to solve this issue...

Co-authored-by: NathanVss <nathan.vasse@gmail.com>