🚨(frontend) allow node import resolver to match cjs files
By default, common js files with extension .cjs are not detected by eslint node
resolver/importer. On our side, we are using glob in node environment so we are
using its common js version so in order eslint is able to resolve the import of
this package we have to update our eslint configuration.