🌐(frontend) change label on CourseProductItem
Currently, when a user has purchased a product, we display a label "Enrolled"
in place of the product price. But this label can be confusing, as the user has
to enroll to course runs for each target courses composing the product. So to
prevent confusion, we decide to replace "Enrolled" label by "Purchased".