✅(frontend) fix flaky tests
The test `CourseProductCourseRuns.renders enrollment not started` was flaky due
to the manner we changed the startDate of the course run to set it in 2 months
relative to today.

The test `CourseProductItem.renders product information` was flaky when a
product used XOF currency because this currency contains a NNBSP in its
representation (F FCA).