✨(frontend) add useOrdersEnrollments
We need an infinite list of orders and enrollments, merged together.
Based on a decision taken in meeting we decided to not implement this
on the backend side. So we needed to create a dedicated algorithm on
the frontend side that re-created this merged list. That's the goal of
this new (complicated at first glance) hook.