✨(frontend) add computeState for course runs
As we are moving course runs list from DjangoCMS to React we need to be able to
compute in real time the state of course runs to resolve caching issues that we
had: as DjangoCMS renders cached pages, even if course runs dates relative to
the current date should cause state to be different it wasn't re-computed on
the DjangoCMS side. By re-computing it on the frontend side we ensure that
their state is always up-to-date.

Resolve #1663