🏷️(frontend) make CourseRunEnrollment use different type
As of now we need to implement the course run list in React, we need to align
CourseRun types accross the component that can appear in this list. Previously
CourseRunEnrollment was using a custom based CourseRun type that wasn't a
faithful representation of the course runs. Especially for the state attribute.