✨(frontend) display enrollment state for ongoing course run
When a course that you've enrolled in reaches the end of its enrollment date,
the button to "Go to course" disappears. This confuses users into thinking that
they do not have access to the course anymore. So now course runs are all
rendered through React, we are able to easily retrieve enrollment state for each
 of them. In order to fix this issue, we retrieve enrollment state for ongoing
 course runs.

Resolve #1981