✨(frontend) create ProductCertificateFooter
There a feature that allow the user to purchase a product certificate
from a enrollment item in the learner dashboard.
If a product have been bought and the course is passed, the user can
download his certificate in this same page.

Futher developements will adapt the sale tunnel design / wording for product
certificate.
Both purchase and download are working with this commit.