🩹(frontend) fix CourseGlimpse regression
Now CourseGlimpse is used with two kind of resources, joanie Course and richie
Course. In order to that, we created a little util that map both resource
properties to a common interface. Unfortunately, we forgot to map some
properties of the Richie course interface.

Furthermore, we rework the CourseGlimpseFooter component props. Now it takes
only course state as props instead of a full course object. But in our factory,
we are overriding this component and we need to access the full course object.