🗃️(backend/core) add `display_mode` field to `CourseRun` model
Editor users wants to be able to customize the look'n feel of Joanie Product
widget. To take this need, we add a `display_mode` choice field to `CourseRun`
model then bind this new field to the React widget props serializer. In this
way, frontend will be aware in which mode joanie product should be displayed.