🚸(frontend) lighten SaleTunnelStepValidation information
According to marketing, at the step validation too much information are showed
to the learner just before the payment that is weird. In order to improve this
we decide to hide target course detail by default. Then the learner is able to
click on the target course title to show its details.