🐛(templates) fix templates using course-glimpse-list block
The course-glimpse-list has been recently changed in order to use it within
teacher dashboard. We did not reflect this update in our templates using
previously this class so the layout was broken. We update templates to use the
new structure of course-glimpse-list which now requires to wrap title within a
header block and content in a content block.