📌(deps) set @testing-library/dom as resolution
The renovate PR made the tests fail due to lots of new warnings.
This kind of warning: update not wrapped inside act ...
They were mainly throwed from user event API calls, but internally
this library already calls act() so it is not supposed to happen.
In reality it happens when different versions of @testing-library/dom
are used by the framework wrapper and the main one loaded by our
project.

See testing-library/user-event#906