🐛(frontend) do not handle error on validated state in OrderStateMessage
The OrderStateMessage does not handle validated state properly. Indeed, this
component does not have message to display for this state and, it handles an
error like this state was unknown. Of course, we only want to handle error in
case an unknown state is provided but not when the provided state does not have
message to display.