🐛(frontend) fix old ways of guessing signature
Previously the way to verify if an order needed a signature was just to
check if it had an attached contract and verifying that it isn't signed.
But in reality the real way to determine it is to verify additionnaly if
the product has a contract_definition, where an order could need a
signature even if its contract is null.