👷(circleci) publish python package on every merge on master
We use setuptools_scm to automatically generate the release version of
richie. When the current commit is a tag, the version tag will be used.
Otherwise it guess the next version number with unique info like the
current commit hash. The CI is refactor to take advantage of this
feature.