♻️(frontend) remove ContractFrame.product
product is used to check its contract_definition.
It was used to know if an order is signed or not.
At this point in the process, we can assume that ContratFrame would not
be opened without a contract definition.

In the incoming contract listing page, we'll not have this product data.