♻️(frontend) remove CourseProductProvider
When introducing SaleTunnelContextProvider to ensure o/c principle
inside the SaleTunnel we realized that CourseProductProvider was
not needed anymore and could be removed in favor of the context of
SaleTunnel, making things more straightforward. We also anticipate
the future of the Sale Tunnel which will get more and more complex.