🐛(frontend) fix various cache order issues
We had a bug where the cache invalidation triggered when a contract
is signed was not impacting useOrdersEnrollments. This was causing
the "Sign" button to be still displayed. A related bug where after
buying a product from the syllabus page, the newly created order
was not shown was also fixed by theses changes.

Fix #2157