💄(frontend) hide order sidebar, hide filters space
We decided to simply hide the sidebar inside the order details because
it was empty, and in the future we will just have to remove the display
none clause to show it again. Also remove the empty space caused by
the filters container, now we only show it if there are filters.