🩹(backend) properly set AUTHENTICATION_DELEGATION.PROFILE_URLS
Now that we have a learner and teacher dashboard, we want to replace the lms
dashboard route by the richie dashboard routes when Joanie is enabled and the
feature flag REACT_DASHBOARD is set to True. Unfortunately, currently, the
current logic is not working as it needs some information set during
configuration post_setup.