💄(frontend) fix organization logo link size issue
Currently, the organization logo could be truncated if its dimensions does not
match the aspect-ratio defined in style. E.g a 1/1 logo was truncated as the
aspect ratio is 2/1.

Furthermore, joanie is returning a srcset property into its file detail, so we
can take benefit of this attribute to properly set organization logo size in
responsive context.