🗑️(frontend) hide contracts menu in course views
Currently in teacher dashboard, we are displaying contract in courses and
organizations views. For now, we decide to show this menu only on an
organization page. Indeed, display contracts within a "simple" course (not a
training) has no meaning. About trainings, we have to rework to manner to get
contracts with right filters and ensure that we are able to add signature
feature on those views. That's why we decide to hide those menus for now and
once we will have setup download and signature on organization view we will work
 to make it work on course views.