🩹(frontend) clean billing address before order submit
It appears there was a typing issue that lead to pass a wrong payload. This
could break payment logic when working the dummy payment backend. Indeed, the
billing address should not contain id and is_main fields but currently the
address was not cleaned up before putting it in the payload. We now use the new
`ObjectHelper.omit` utils to remove id and is_main fields from the selected
address.