♻️(frontend) create Organization and Learner ContractFrame
We have created a component ContractFrame which embeds all the logic to sign a
contract from an invitation link but as first this component works only for
learner. We now have to implement it for organization. That's why we first
abstract the ContractFrame component then create two components which
"implements" the new AbstractContractFrame component, LearnerContractFrame and
OrganizationContractFrame