🚚(frontend) rename DashboardContractLoaders to DashboardContractLayout
In react router dom, the term loader aims to represent a container component in
charge of fetching data and populate presentational children. It was relevant
for the DashboardContract page but now, data fetching has been moved into
children component so the Loader is no more relevant that's why we replace this
term by Layout.