🐛(frontend) fix getCourseGlimpseProps intl
In some case, the `getCourseGlimpseProps` method needs to use an intl interface
to compute the course_url prop. Currently, we were creating the intl object
inside this method but we did not provide messages so this was not working. In
order to fix that, we are now passing an intl object to this method instead of
locale to afterward create an intl object.