✨(frontend) display course link in order detail view
Currently, in the order detail view, if a user is enrolled to a course run we
only display a labelled to tell it, it is enrolled but there is no link to
access to the course that is weird. So we revamp this UI to add this link and
keep information that the user is enrolled.

Resolve #2232