🐛(frontend) fix logout issue
We realized that, on logout, setting react query ["user"] to null was
taking too much time to be persisted in SessionStorage resulting in a
redirection to the homepage before the data being flushed into the
SessionStorage. (It is caused by throttleTime of the persister )
Sometimes the user was not able to logout due to this
issue.

Co-Authored-By: jbpenrath <jb.penrath@gmail.com>