🐛(frontend) fix teacher courses list scroll top
We had a bug where the courses list scroll was reset at the top of the
page.
useUnionResource isLoading is switching between true and false each time
it load a new chunk of results.
As isLoading was used in TeacherDashboardCourseList to switch display
between the spinner and the courses list, the page scroll was sometime
reset when the list disappear.

Resolve #2227