🐛(frontend) fix future course access link
In Learner dashboard when a course isn't yet open, we displayed a
"access course" link that was broken.
The LMS return a 500 for theses courses and the user arrived a blank
page.

There we decide to hide theses link and we use this change to upgrade
the period label of the enrollment.