🐛(frontend) fix SaleTunnel several cache issues
We encountered several issues in specific use cases with the SaleTunnel:

- Open and close the payment interface, when trying to open again the
frontend tries to create a new order, but there is already an existing
order: we should simply re-submit it.
Fixed by: invalidate cache after aborting.

- Open and close payment interface and SaleTunnel, the PurchaseButton
is not displayed: it should because the order is pending.
Fixed by: invalidate cache after aborting.

- If we open a new tab and there is already an existing pending order,
the last step of the SaleTunnel (resume) crashes because the order
variable it uses is not defined.
Fixed by: Bring up useProductOrder to SaleTunnel in order to make the
existing order available accross the context.